requirements.txt
# Core LangChain framework
langchain>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0

# Vector stores
chromadb>=0.5.0
faiss-cpu>=1.7.4

# PDF parsing
PyMuPDF>=1.24.0

# Embeddings (local HuggingFace)
sentence-transformers>=3.0.0

# Reranking
# (sentence-transformers includes cross-encoder support)

# Web UI
streamlit>=1.38.0

# Environment management
python-dotenv>=1.0.0

# Evaluation
ragas>=0.2.0

# Utilities
pandas>=2.0.0
